heap: Fix outer scope for GC tracing
So far, the UMA metrics reported for the GC cycle (e.g. V8.GCScavenger,
V8.GCCompactor, V8.GCFinalizeMC) report time spent during two successive
epochs: the finalizing part of the last epoch, as well as time spent
during the new epoch. This CL, which is part of the migration so that GC
statistics are reported directly to the Recorder API, implements the
same behavior for V8.GC.Cycle.* metrics via GCTracer scopes, ensuring
that each part is associated with the correct epoch.

Bug: chromium:1154636
Change-Id: I14b11f4e954adbb2e022a53f0585a29df052c5bf