[compiler] Fix another SIMD register aliasing issue
If a SIMD operation has a fixed output register (in this case q0), but
one of the two aliasing FP registers is in use (in this case d1), then
this FP register needs to be spilled.

This is similar to issue https://crrev.com/c/3347576, but in this case
{ChooseRegisterFor} is not used, because the output is at a fixed
register.

R=mslekova@chromium.org

Bug: chromium:1282197, v8:12330
Change-Id: Iea85222b2d8bfb9baf50231f0e866b1354fb761c