[runtime] Update SavedClassVariable in ScopeInfo
ScopeInfo will contain either inlined (array) local names or
a hash table (names => index) containing the local names.

If we have the local names inlined, we should save the class
variable index, since we can access the name in O(1).

If we have a hash table instead, we should save the class
variable name, since we can access the index in O(1).

Bug: v8:12315
Change-Id: Ifd9ae4f285d11fc034e8560c8558038b38a474fb