Add prototype of WASM Trace proposal
Implements an intial prototype of the WASM Trace proposal. A custom
section containing offsets to functions is decoded into trace
instructions that are inserted into the function. In Liftoff, these
are directly inserted. In TurboFan, these are added as StackEffect's,
this is a work in progress.

Traces will only be decoded and added when a flag is given to V8,
currently "--experimental-wasm-instruction-tracing". If a trace is ever
not valid or an error occurs, it is safe to just throw them away.


WASM Trace Proposal:
https://github.com/WebAssembly/instrument-tracing

Code Annotations Tool Convention:
https://github.com/WebAssembly/tool-conventions/blob/main/CodeAnnotations.md

Design Doc:
https://docs.google.com/document/d/1739a_LXbavBnek7pa0uqhHOCz8IJ56mn2C2Yvbssvkg/edit?usp=sharing

Bug: 1090122, 1252113
Change-Id: Id4690d8deca482ff0e863761668ffabca159bd29