[base] Fix AllocateInternal on Fuchsia when VMAR offset is zero
In that case, the previous logic would incorrectly conclude that no
target address was specified, and would potentially place the allocation
elsewhere in memory, not at the desired address. This CL now passes both
the target address and the VMAR base address to AllocateInternal, which
can then correctly determine whether a target address was supplied.

Bug: v8:10391
Change-Id: Ia8215440a790b4a2a0c8d33f623d3ecb6a731a97