[debugger] Pass break reason for interrupt along with BreakRightNow
This explicitly passes along the break reason when requesting a pause
that is handled via an interrupt. Pushing the break reason is not
enough (as done before), as the break reasons that are pushed
are only consumed if we are making use of setPauseOnNextFunctionCall,
and not an interrupt.

Fixed: chromium:1292519
Change-Id: If8635c9397a7b9a1e6757be8048c9edc613c27f8