handles: Only reclaim TracedNode during atomic pause
TracedNode is used as backing node for v8::TracedGlobal (with
destructor) and v8::TracedReference (no destructor).

A future CL adds concurrent marking for TracedReference which requires
that the node stays around until the atomic pause to avoid
synchronization with the concurrent marker.

This CL prepares TracedNode in prolonging the lifetime until the it is
actively cleared ("sweeped") in the atomic pause. This allows for
spuriously keeping alive a TracedNode for an additional GC cycle in
the case the the node was destroyed while the marker is running.

Bug: v8:12600
Change-Id: I9688c83a42b70d352c84613485f37242b1b910a6