[compiler] Change liveness to use a flat array
Bytecode liveness needs a mapping from offset to liveness. This was
previously a hashmap with a very weak hash (the identity function) and
both inserts and lookups showed up as a non-trivial costs during
compilation.

Now, replace the hashmap with a simple flat array of liveness, indexed
by offset, pre-sized to the size of the bytecode. This will have a lot
of empty entries, but will have much better runtime performance and
probably ends up not much lesss memory efficient as a hashmap if the
hashmap has to resize inside the Zone.

Change-Id: Id21375bfcbf0d53b5ed9c41f30cdf7fde66ee699