cppgc: young-gen: Implement GenerationalBarrier with eager tracing
The CL implements the barrier
(HeapConsistency::GenerationalBarrierForSourceObjectWithCustomCallback)
that eagerly processes the source object looking for slots inside it.
To record them, the CL brings the slot pointer into the context of
cppgc::Visitor's functions.

This type of barrier is planned to used for Blink backing stores when a
single inlined object or sufficiently small number of inlined objects is
added.

Bug: chromium:1029379
Change-Id: I7519ac71e422eb593a0a691788d2a9911ef2d13b