[builtins][arm64] Prevent use of mid-tier register allocator for builtins
The mid-tier register allocator does not mark a block as needing a frame
before it spills a value to a stack slot and so accesses below the stack
pointer when used for certain builtins for arm64.

Also prevent use for FOR_TESTING, BYTECODE_HANDLER and C_WASM_ENTRY
so as to not use the mid-tier register allocator at static compilation
time, allowing the mid-tier register allocator to be used for all other
functions and pass all tests.

Bug: chromium:1299793
Change-Id: I74175d75847d092d1d5373661d536a299a7da1b0