heap: Refactor sweeping finalization for young GC
Before a young GC, the scavenger finalizes sweeping in case there
are no more running sweeping jobs, to avoid unnecessarily pausing
and then resuming. This CL moves this sweeping finalization from

ScavengerCollector::CollectGarbage to Heap::CompleteSweepingYoung
so that it is also performed for the minor mark-compactor and it
is correctly attributed to the previous full GC cycle, and not to
the beginning young cycle. Furtermore, it finalizes CppGC sweeping
if there are no more running sweeping jobs.

Bug: chromium:1154636
Change-Id: Ic9ba4803f49db32c0a539f080329f012859bc8bc