Reland "[fuchsia] Use SDK provided gn templates for v8_unittests"
This is a reland of commit 8d51d561d5186b483d904e95b56549897683017e

Further digging found that including cmx shards doesn't work OOT.
We can revisit when migrating to CFv2, but until then each cmx
file needs to be self-contained. We had actually discussed this
approach in the initial implementation because the previous approach
was including a lot of things that v8_unittests doesn't actually
need.

Original change's description:
> [fuchsia] Use SDK provided gn templates for v8_unittests
>
> cr_fuchsia_package is deprecated in favor of using the Fuchsia
> SDK provided rules directly.
>
> Bug: chromium:1092804
> Change-Id: I86a59b6a717cb0aa8c3473e8410bdd98f0ffd042
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3537883
> Reviewed-by: Victor Gomes <victorgomes@chromium.org>
> Reviewed-by: Wez <wez@chromium.org>
> Reviewed-by: Camillo Bruni <cbruni@chromium.org>
> Commit-Queue: Bryant Chandler <bryantchandler@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#79763}

Bug: chromium:1092804, chromium:1310982
Change-Id: I2ce8bd11cbaab9e4229bf9f255befe000ddb324c