[wasm] Deduplicate continuation-resuming callback
For stack-switching, we create a callable object from the
WasmResume builtin and pass that as the onFulfilled argument
of Promise#then. We don't need to create this callable object each time
we suspend. Instead, create it when we initialize the Suspender object
and store it there.

R=jkummerow@chromium.org
CC=fgm@chromium.org

Bug: v8:12191
Change-Id: If8495493a71794cddc81b21a17a821fed8f4ede7