Merged: [valueserializer] Implement a compatibility mode for deserializing broken data
For a while, we shipped a version which writes version 13 data with
JSArrayBufferView flags, and then fixed version 13 to not include the
flags.

This CL adds a compatibility mode for parsing the the version 13
data which includes the flags, since it still occurs in the wild.

Bug: chromium:1314833,chromium:1284506
(cherry picked from commit 682ba0efd3ad85a586a6958e56c6a944d277ccf8)

Change-Id: I82306a9efa1f1121da6941f6afb8f1b0756f0fe0