[heap] Split NewSpace to base and concrete classes
This CL is the first step towards creating a paged new sapce.
The CL creates a base class for new space that holds all fields not
specific to the semi space based implementation, and methods that do not
interact with the semi spaces. Methods are moved as is to the new base
class.

Future CLs will rename the classes, split/refactor additional methods,
and make other methods virtual.

Bug: v8:12612
Change-Id: Ibd8ec5135d66daf0fd025493bfaff537d61e049f